[SC-44129]Add LogStore.readAsIterator
This PR proposes a new method `LogStore.readAsIterator` to load the file on demand.

`LogStore.readAsIterator` returns `ClosableIterator` which can be implemented by a LogStore implementation to load the content on demand. We also implement a new class `LineClosableIterator` which can turn a `Reader` to `ClosableIterator`. This class makes it easy to implement `readAsIterator` for a LogStore implementation.

This PR also uses this API in `describe history` to solve its scalability issue.

The new unit tests.

Author: Shixiong Zhu <zsxwing@gmail.com>

GitOrigin-RevId: fb136c38c8070c4a822ac667063710b1bfc1f02c