Delta Standalone Reader (#51)
In this PR, we add a new project `standalone` which implements a Delta reader without using Spark. The reader provides APIs to access Delta transaction logs programmingly and a simple interface to read the data files in the table as well.

We also update the Hive connector project to use this new project and get rid of the embedded Spark from the Hive connector.