[SC-57630][DELTA] Handle struct schema evolution for MERGE INTO.
## What changes were proposed in this pull request?

Allow schema evolution to work for for nested columns, including proper by-name resolution.

The finalSchema field in DeltaMergeInto is removed because it doesn't end up being necessary. We added it in the previous schema evolution PR to handle calculating the final schema in deltaMerge.scala, but that calculation is now deferred to PreprocessTableMerge where it can be added directly to the merge command without intermediate steps.

## How was this patch tested?

new unit tests

Author: Jose Torres <joseph.torres@databricks.com>

#13248 is resolved by jose-torres/mergeupdate2.

GitOrigin-RevId: b8fbeb5a38005df6282eb50bf7cec795b7bccbdd