[SC-65167] Check RemoveFile when OptimisticTransaction.readWholeTable is called
When a conflict commit contains RemoveFiles, we should check whether it removes files read by OptimisticTransaction. Currently, we only check whether RemoveFiles contain any of files in `readFiles`. However, `OptimisticTransaction.readWholeTable` doesn't update `readFiles`, so this is not enough when `OptimisticTransaction.readWholeTable` is called.

This PR disallows any RemoveFiles when `OptimisticTransaction.readWholeTable` is called so that we are able to detect this case.

Impacts:
- Reading a Delta table and writing to the same table using DeltaSink

Author: Shixiong Zhu <zsxwing@gmail.com>

GitOrigin-RevId: 6158df82baf5a8bfca288db343eeaa6fc7cb9165