[SC-65096][DELTA] Pass schema evolved output into the LogicalRelation of the target plan in MERGE INTO.
## What changes were proposed in this pull request?

Pass schema evolved output into the LogicalRelation of the target plan in MERGE INTO. Right now we replace the attribute reference types only in a projection *above* the LogicalRelation, which isn't valid - it means the same exprId is used for both the pre and post-evolution datatypes, causing inconsistent codegen depending on which type happens to be picked up first. After this PR, the exprId should always have the post-evolution datatype.

(Ideally, we wouldn't need to do this kind of surgery at all, but I don't see another way to resolve this without refactoring the entire MERGE implementation.)

## How was this patch tested?

The primary test is LogicalPlanIntegrity.checkIfExprIdsAreGloballyUnique, which checks that all references to a particular exprId have the same type. This check is disabled in master pending resolution of SC-67287 but I verified locally that this change makes the check pass in MergeIntoSQLSuite. This also fixes the repartition(1) repro in SC-65096, along with the original issue we saw in CDC (but re-enabling the CDC schema evolution we should defer to a separate PR).

Author: Jose Torres <joseph.torres@databricks.com>

#17249 is resolved by jose-torres/fixbehavior.

GitOrigin-RevId: 7bf4230699a8d50e6400f6a4fa44880d80401d32