[SC-69350][DELTA] Do not store write options in the catalog for Delta
Write options such as `replaceWhere` and `mergeSchema` can be stored in the transaction log, as well as the catalog when using the DataFrameWriter with `saveAsTable`. This has been a bug, as write options should not be stored in the transaction log. Nor do we need to store anything in the Catalog for Delta. This PR cleans up these properties from the catalog as well as the transaction log.

However, there may be users who depend on this behavior. Therefore we do two things:
  1. Introduce a legacy flag so that users can revert to the old behavior
  2. We continue to store any Delta specific configurations, which are prefixed by `delta.` to the transaction log

Unit test

Author: Burak Yavuz <brkyvz@gmail.com>

GitOrigin-RevId: 3bf8db1b25f94096b13855a806a24122871a1585