[SC-63766] Add insertionTime column to delta log
We introduce insertion time tag to the delta log to indicate when the data in files has been originally inserted. This timestamp should be preserved during the lifetime of the data.
After merging a group of files, the insertion time of the new file is set to the latest of the insertion times of the original files, unlike the modificationTime which depicts the time the file was written.

new unit test

Author: Sabir Akhadov <sabir.akhadov@databricks.com>

GitOrigin-RevId: 7b2bd87b7a17e3cedc3c97c1724304331729f214