[SC-70676] Deleting delta table directory before first DF is defined leads to empty result
This PR fixes the following corner case:
```
spark.range(1,5).write.format("delta").save("/tmp/t1")
// Delete directory /tmp/t1 in filesystem
spark.read.format("delta").load("/tmp/t1") -> unexpected empty result.
```
In this case we will get an InitialSnapshot in DeltaLog.createRelation for a
read query. Before this PR we proceed and produce empty result, which is not
desired behavior, as an empty Delta table still need to have the directory and
some metadata in it. With this PR we will now error out for this case.

Author: Meng Tong <meng.tong@databricks.com>

GitOrigin-RevId: a296587aa1789b2c9d4be621ed1558213dd95e37