[SC-70791]Handle null values for constraint of generated column
The constraint of generated column is using `EqualTo` which always returns `null` when any child is `null`. This means if the provided value and the calculated value are both `null`, it won't pass the constraint.

This PR changes the expression to `EqualNullSafe` to fix the issue. It also addresses the outstanding comment in #18182

The new unit test

Author: Shixiong Zhu <zsxwing@gmail.com>

GitOrigin-RevId: 3538b0a66251ca903f98e21ac35afc67f24a9e29