[SC-65677] Avoid checkpointing the same version
Before this PR, when we create a checkpoint, we use the snapshot established at the transaction start and create a checkpoint for that version. This might cause multiple transactions to checkpoint a same version and potentially lead to corrupted checkpoint status. In this PR, we change to checkpoint the version committed by this transaction to avoid such scenario. The downside of this approach is that we pay extra cost to get the new snapshot occasionally. We expect this is a rare case and we can tolerate the cost.

Existing tests.

Author: Meng Tong <meng.tong@databricks.com>

GitOrigin-RevId: 1858f58f69e0618924709631b67f81fe1d0b863d