[SC-74152] Disallow column type change in Generated Columns
Currently Delta allows some safe type change such as from SMALLINT to INT.

However, this may break generated column contract. For example, let's say we have a column c1, and a generated column c2 defined as `CAST(hash(c1 + 32767s) AS SMALLINT)`. When c1's type is SMALLINT and we insert `32767s`, the expression will return 31349, but if c1's type is INT and we insert `32767`, the expression will return 9876. This means changing the column type may require to rewrite the existing data. But since it's too heavy, we can simply disallow it.

New unit test

Author: Shixiong Zhu <zsxwing@gmail.com>

GitOrigin-RevId: 93a0475fed83ec6a751e4b05902aec3fa71410a5