[SC-70743] Add ability to incrementally list from a particular file in DeltaFileOperations
Added two methods, `recursiveListFrom` and `localListFrom`, into `DeltaFileOperations`.

Currently, the only way to list files is to specify a directory path, then all files under that directory will be listed. This is wasteful if we only want _new_ files after a certain filename, instead of needing to re-list the entire directory and filter thereafter.

These two methods allow you to specify a directory name and a path (in that directory) from which to list from. Then, taking advantage of `LogStore.listFrom`, only files with filenames after the specified path will be returned.

- Added tests in `DeltaFileOperationsSuite`.

Author: Howard Xiao <howard.xiao@databricks.com>

GitOrigin-RevId: af5712e60224835cf00e247c6a9f740e04126968