[SC-59632][DELTA] Add a test for MERGE INTO schema evolution for unlimited clauses
Schema evolution didn't originally work with unlimited clauses. We need a test to prevent this from regressing in the future.

Right now there's just a Scala suite test, since the unlimited clauses test harness doesn't support evolution and the evolution test harness doesn't support unlimited clauses, so it's complicated to write a test that the ACL and CDC extensions of the SQL merge suite will correctly be able to skip.

new unit test

Author: Jose Torres <joseph.torres@databricks.com>

GitOrigin-RevId: 4e492459966fafb23d1d5b3d4fea95656d02cf55