[SC-74853] Convert Delta OSS into a multi module SBT project
Making Delta a multi-module project will enable us to add other sub-modules. For example, we can then add a contribs sub-module that can have contributions from the community that needs to be very closely tied to the delta-core project (hence in this repo, and not delta/connectors) but does not have the same level of maturity as delta-core.

Changes made in the Delta repeeo
- Moved all files in root/src/ to root/core/src/
- update build.sbt to multiple modules
  - Removed dependency on spark-packages.

existing tests

Closes https://github.com/delta-io/delta/pull/644

Author: Tathagata Das <tathagata.das1565@gmail.com>

GitOrigin-RevId: 68038d27302e82f6e680fe717633109757e48ba0