[SC-76061] Add better error message for concurrent write failures on empty path
When two concurrent writers write to an empty path for the first time, the 2nd writer would fail because of ProtocolChangedException. In particular, the first writer would have updated the reader and writer version of the protocol and the second writer (even though could be updating the same reader-writer version) would consider this as a transaction conflict and throw a ProtocolChangedException error message.

This change adds an additional information to the error message that this was due to concurrent transactions writing to the same empty path.

Author: Vijayan Prabhakaran <vijayan.prabhakaran@databricks.com>

GitOrigin-RevId: 96e85c3363fd04181f7880e5907457b14b35b197