[SC-18397] LogStore public API
This PR adds the new LogStore API designed to be a public API. The existing LogStore API is not changed. The new API is based on the existing API but we cleaned up a few places to make it simpler and easier to work with. As the new API and old API are not binary compatible, this PR also adds `LogStoreAdaptor` which adapts all implementations of new API to existing API. All implementations of existing LogStore API will continue to work as they are today without any changes.

Author: Meng Tong <meng.tong@databricks.com>

GitOrigin-RevId: b389717c28fed9ea4ffbfdd999dc4265615270af