[SC-71820] Support Generated Columns in update command
Currently, when a source column of a generated column is updated in UPDATE command, we don't update the generated column accordingly. For example, given a table that has a generated column `g` defined as `c1 + 10`. For the following update command:

```
UPDATE target SET c1 = 100
```

We will copy the old value of `g` . This is not correct, and it will fail because of the constraints we apply for generated columns. The correct approach should be updating `g` to `110` according to the new value of `c1`.

This PR updates `PreprocessTableUpdate` to generate update expressions for generated columns automatically when the user doesn't provide update expressions.

New unit tests

Author: Shixiong Zhu <zsxwing@gmail.com>

GitOrigin-RevId: 5b6d3c5d37439d18b158269c2165e1b12c068209