[DELTA-OSS-EXTERNAL] Add null check to avoid NPE when Exception message is null while merging.
Hi,

When we interrupt a stream that updates a Delta table using _merge_ mode, the JVM runtime throws a plain `InterruptedException` with no description and the `AnalysisHelper` expects that all exceptions have a message and this causes a `NullPointerException` on such cases. This PR adds a null check to avoid that. Also added a unit test for this scenario.

Thank you,
Bruno

Closes delta-io/delta#648

Signed-off-by: Shixiong Zhu <zsxwing@gmail.com>

Author: Bruno Palos <brunopalos@gmail.com>
Author: Shixiong Zhu <zsxwing@gmail.com>

#21879 is resolved by zsxwing/pouscxop.

GitOrigin-RevId: 6be7561f24dd6b26a8f8732e82e25896234071a8