[SC-76446][DELTA] Set isBlindAppend correctly for DSv2 plans.
ActiveOptimisticTransactionRule is not idempotent, but V2 write plans will be planned again based on the optimized version of the original plan through the V1 fallback paths. So we have to skip these rules for V2 write plans, or the file indices will be pre-pinned and thus not invoke the proper logic to signal their scans to the transaction.

new unit test

Author: Jose Torres <joseph.torres@databricks.com>

GitOrigin-RevId: 51fc053d022ebcf6b45bf4441eed799d7d68d969