[SC-78050][DELTA] Call `deltaLog.update()` in DataFrame read path
## What changes were proposed in this pull request?

When a snapshot was created as an `InitialSnapshot` for a Delta table and is cached as such (for example, a race condition due to unmounting and mounting paths), then all following reads on that Delta table would return a “This path is not a Delta table” error. This PR adds an `update()` call to the Dataframe read path to prevent this from happening (and give the valid table).

This is done by forcing the computation of `snapshot` when we create a `BaseRelation` for `DeltaTableV2`. In short, this will call `deltaLog.update()` so we ensure that the check whether or not the table exists is accurate. This costs an additional RPC but is deemed necessary for correctness.

## How was this patch tested?

Added a unit test to simulate reading from a table with cached `InitialSnapshot` and a valid DeltaLog.

Author: Zach Schuermann <zach.schuermann@databricks.com>

#23778 is resolved by schuermannator/sc-78050.

GitOrigin-RevId: 8fd732bbf39788f92ea390f720aa9bb4246e8d12