[SC-74210] Drop NullType columns for SQL read path
NullType column is not very useful as they do not contain any contents. Hence we used to
drop this NullType column when we create a table from DataFrameReader, but we did not
do the same thing on SQL read path. This PR unifies the behavior, which will drop NullType column
always in any read/table/sql APIs.

Unit tests testing different read APIs.

Author: Junyong Lee <junyong.lee@databricks.com>

GitOrigin-RevId: 9b55e8fb5e51ffbfb86832a811668e9b920c225e