[SC-70216] Support generated columns in merge command
## What changes were proposed in this pull request?

This PR adds support for generated command in MERGE ... UPDATE case. Previously, if the generated column is not explicitly updated, we will copy over the old values, which would potentially break the generated column check constraint and fail the query. With this PR, the values of generated columns will be computed correctly using the (potentially updated) referenced columns.

This PR mostly reuses the utility functions for UPDATE to generate the correct update expressions, with some small changes needed to cover the schema evolution case.

## How was this patch tested?

Added new unit test.

Author: Meng Tong <meng.tong@databricks.com>

#23499 is resolved by mengtong-db/generated-column-merge.

GitOrigin-RevId: 6245c07c323255eb4a0db88150e520ef24e02af8