[DELTA-OSS-EXTERNAL] Relax importlib_metadata version required
The current requirements for `importlib_metadata` is too strict (`importlib_metadata>=3.10.0`) and has led to it conflicting with versions required by other python libraries such as `apache-airflow` (`importlib_metadata==1.7.0`). Given that the only feature used from `importlib_metadata` is the `version()` function, and the behaviour of this function has not changed from version `1.0.0`, it is safe to relax the version required to improve compatibility with other libraries.

Thank you.

Changes
- Relax `importlib_metadata` version required from `>=3.10.0` to `>=1.0.0`

Closes delta-io/delta#709

Signed-off-by: Scott Sandre <scott.sandre@databricks.com>

Author: YuXuan Tay <wyextay@gmail.com>

#25854 is resolved by scottsand-db/b5cu40hu.

GitOrigin-RevId: 20976a9ce0b7a321134cc53e5b271d4c98e2ed2c