Fix AnalysisException on multiple insert clauses in merge
With schema evolution on, the following merge query fails with:
```
Error in SQL statement: AnalysisException: There is a conflict from these SET columns: [`c`, `c`].
```
For tables:
```scala
spark.range(2).selectExpr("id a", "id b").write.format("delta").saveAsTable("t")
spark.range(10).selectExpr("id a", "-1 c").write.format("delta").saveAsTable("s")
```
```sql
MERGE INTO t USING s ON t.a = s.a
  WHEN MATCHED AND s.a = 0 THEN UPDATE SET t.b = s1.c
  WHEN NOT MATCHED AND s.a = 7 THEN INSERT *
  WHEN NOT MATCHED AND s.a = 8 THEN INSERT *
```

This PR simply adds a `.distinct` call during `PreprocessTableMerge` when processing matched clauses; then when we calculate the columns that are present in the insert clauses but not present in the target/update, we aren't creating duplicates (which caused the above error).

The above scenario is replicated in a test in `MergeIntoSuiteBase`.

Author: Zach Schuermann <zach.schuermann@databricks.com>

GitOrigin-RevId: 4ca1ade59953221611a560c5d6b00bf047720303