[SC-85374] Should remove Delta column mapping metadata in HadoopFsRelation
Delta column mapping metadata should be dropped before the table schema is passed on to downstream, because such metadata is not needed in Spark SQL and also query plan transformation is not very robust in terms of propagating metadata correctly. Specifically, we fix the place when creating HadoopFsRelation and DeltaTableV2.

New unit tests

Author: Liwen Sun <liwen.sun@databricks.com>

GitOrigin-RevId: 707681a4f7023aa72d82f180fbe112ac4ad20379