[DELTA] StateCache now works with DataFrame in addition to Dataset[A].
This PR updates the `StateCache` helper class so that callers can access its underlying RDD as a `Dataset[A]` or as a plain `DataFrame`. The latter potentially allows for lower runtime overhead by avoiding the need to parse rows to scala objects, in case the calling code does not require it.

While we're at it, split the vague and generic `state` method of `Snapshot` into `stateDS` (a `Dataset[SingleAction]`) and `stateDF` (a `DataFrame`), and update some of Snapshot's internal use sites to use the latter since they were already treating the `state` as a `DataFrame`. For now, all other use sites stick with `stateDS` (= unchanged behavior).

Existing unit tests exercise this code.

Author: Ryan Johnson <ryan.johnson@databricks.com>

GitOrigin-RevId: e8f798fed53422e6fac4fd0bd2ef2efe1067eb6a