[SC-86513][DELTA]Unify Delta code to use the same API to create Hadoop configuration
This PR adds a new method `DeltaLog.newDeltaHadoopConf` and refactors Delta code to use this method to create Hadoop configuration. This is the prerequisite for supporting passing Hadoop file system configurations using DataFrameReader/writer options. Here is the list of the major changes in the PR:

- Add `DeltaLog.newDeltaHadoopConf`.
- Add a scala style rule to OSS Delta to ban `sessionState.newHadoopConf`.
- Add a new method that accepts Hadoop Configuration for each method in log store

Here is what I did:

- Add a new hadoop configuration parameter to all LogStore methods. In other words, we add new methods and remove the old methods.
- Fix all compiler errors and pass all tests.
- Add back the old methods to maintain the compatibility but none of Delta code will call these old method.

Existing tests. This PR should not change any behaviors in the code.

Author: Shixiong Zhu <zsxwing@gmail.com>

GitOrigin-RevId: a9f61e060fa2588621b26be72725da7b0fc4d92f