[SC-85245][DELTA] Refactor logSegment in snapshot management
Loading a log segment in snapshot management consists of two code paths, while most of logics share the common path. This PR refactors them to go a common path, with slightly different options for each. By doing so we expect any future improvement for this path would be easy to maintain and modify.

GitOrigin-RevId: 5dc9fd0f6aa387f36dbaf147cdcf9ded6060cd9a