Reject empty string for string partition column
This PR adds a projection to convert empty string to null values before `DeltaInvariantCheckExec` so that any constraint (NOT NULL or CHECK) can correctly treat empty string as null values.

The change is behind an internal conf which is by default off.

Added new tests to verify that empty strings are correctly rejected when the constraint does not allow null values.

GitOrigin-RevId: af2923e7e5296201f4b3329673f2d87921f0d720