Add null-check for data reader partition field (#219)
Our `RowRecord` APIs say that we will throw a `NullPointerException if field is not nullable and {@code null} data value read`. We are doing this for data fields, but we forgot to do this for partition fields.

We can't test this as we can't insert a row with a null value for a a non-nullable collumn.