[SC-89410] Capture non-fatal exceptions while writing checkponts
We discovered that we might throw an exception to the user when a checkpoint fails during a write command. However, this is undesirable, since the write has succeeded in the user's perspective, even if the checkpoint failed. Capture non-fatal exception types (don't try to capture fatal exceptions like JVM failures though), and log when this occurs

Updated existing unit tests that captured exceptions, and added a new one to test for different types of exceptions (like fs exceptions)

GitOrigin-RevId: 4911f90d5d46bac44defe561bf32ce817f50656c