Add an internal flag for checkpoint to throw exceptions
It's important for our existing test suite to catch any checkpoint related errors. We introduced a flag to control this in case a certain tests (that is not controlled by the original utils.IsTesting flag) to throw exception on checkpoints.

Tested with unit tests, integration test to catch a buggy case

GitOrigin-RevId: 0f2113d219f837bedf4b44e525fd2e8024ba3408