Change the interface of loadAction to reduce unnecessary conversion
In state reconstruction of DeltaLog, we have a state in Dataset[SingleAction] converting it to Dataframe to
massage some columns and sort, then we re-convert back this to Dataset[SingleAction]. This conversion
is not necessary and may have small performance impact, so we improve the interface of loadActions to be
just DataFrame. This will give following benefits: a) Minor performance improvement: conversion from
DF -> DS[SingleAction] -> DF is not needed anymore. b) Since we return DF, the interface becomes more
flexible if any future optimization needs to be applied for sorting and partition purpose.

All existing unit tests (it is a common code path that uses state reconstruction for almost all deltalog operations)

GitOrigin-RevId: 1e1fa5cc9d42ae294c90a0861278619a2a1fdcc3