configure_spark_with_delta_pip fix to stop overwriting spark.jars.packages
This pull request enhances `configure_spark_with_delta_pip` function to add an optional parameter extra_packages to allow for users to add their custom jars as required.

A test was added class was added `PipUtilsCustomJarsTests` to unit test by adding a duplicate of the delta jar and see if spark conf contained both jars in the assertion.

A new test class was added due to having to modify the spark session rather than using the existing test class.

This resolves issue #889.

Closes delta-io/delta#909

Signed-off-by: Shixiong Zhu <zsxwing@gmail.com>
GitOrigin-RevId: e667ef3aa63e8546fe12dd820fed8dc790919abe