Disable special handling of empty transactions in Delta
Delta reduces the isolation level for empty transactions e.g. MERGE/DELETE/Update etc. This causes SERIALIZATION issues when mixed with "blind append special handling" in Delta.

This PR removes the special handling of empty txns behind a config so that we can go back to earlier behaviour (the wrong behaviour) if needed.

GitOrigin-RevId: 190aa3f461f637202322e371a26d8a32d5f1c9af